Simplify imported lib
